+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.1.69 (10/02/2020)
+
+#### 🚀 Enhancements
+
+- [**groovy**][**master**][**windows**] Support dockerLogin step in Windows [#381](https://github.com/elastic/apm-pipeline-library/pull/381)
+
+---
+
 ## v1.1.68 (07/02/2020)
 
 #### 🐛 Bug Fixes
@@ -650,11 +658,11 @@
 
 #### ⚙️ CI
 
+-  feat: add docker images to the packer cache [#152](https://github.com/elastic/apm-pipeline-library/pull/152)
 -  fix: used trustworthy Docker images [#151](https://github.com/elastic/apm-pipeline-library/pull/151)
 -  fix: post step is in the wrong place [#147](https://github.com/elastic/apm-pipeline-library/pull/147)
 -  fix: make the checkout of the repo [#146](https://github.com/elastic/apm-pipeline-library/pull/146)
 -  fix: remove documentation stage and related steps [#145](https://github.com/elastic/apm-pipeline-library/pull/145)
--  feat: add docker images to the packer cache [#152](https://github.com/elastic/apm-pipeline-library/pull/152)
 
 ---
 
@@ -1010,8 +1018,8 @@
 
 #### 🚀 Enhancements
 
--  [APM-CI] add cache to GitHub API REST calls [#42](https://github.com/elastic/apm-pipeline-library/pull/42)
 -  [APM-CI] add token cache to the codecov step [#41](https://github.com/elastic/apm-pipeline-library/pull/41)
+-  [APM-CI] add cache to GitHub API REST calls [#42](https://github.com/elastic/apm-pipeline-library/pull/42)
 
 ---